requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
google-cloud-texttospeech>=2.16.2
google-cloud-language>=2.10.0
google-cloud-speech>=2.25.0
sqlcipher3-wheels>=0.4.0
mangum>=0.17.0
streamlit>=1.28.0
pytest>=7.4.0
structlog>=23.2.0
httpx>=0.25.0
python-multipart>=0.0.6